requirements.txt
# Core dependencies
pymilvus>=2.3.3
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# UI and visualization
gradio>=4.0.0

# NLP and text processing
spacy>=3.7.0
jieba>=0.42.1
scikit-learn>=1.3.0

# Async support
asyncio>=3.4.3

# File handling
pathlib>=1.0.1 